IPRT: GCC 8.2.0 fixes
svn:sync-xref-src-repo-rev: r124129

+++ src/VBox/Runtime/testcase/tstRTHeapOffset.cpp
@@ -1,4 +1,4 @@
-/* $Id: tstRTHeapOffset.cpp 69111 2017-10-17 14:26:02Z knut.osmundsen@oracle.com $ */
+/* $Id: tstRTHeapOffset.cpp 73502 2018-08-05 12:40:50Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT Testcase - Offset Based Heap.
  */
@@ -100,7 +100,7 @@ int main(int argc, char **argv)
         {        16,          0,    NULL,  7 },
     };
     uint32_t i;
-    RTHeapOffsetDump(Heap, (PFNRTHEAPOFFSETPRINTF)RTPrintf); /** @todo Add some detail info output with a signature identical to RTPrintf. */
+    RTHeapOffsetDump(Heap, (PFNRTHEAPOFFSETPRINTF)(uintptr_t)RTPrintf); /** @todo Add some detail info output with a signature identical to RTPrintf. */
     size_t cbBefore = RTHeapOffsetGetFreeSize(Heap);
     static char const s_szFill[] = "01234567890abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
 
@@ -184,7 +184,7 @@ int main(int argc, char **argv)
         RTTestIPrintf(RTTESTLVL_ALWAYS,
                       "Warning: Either we've split out an alignment chunk at the start, or we've got\n"
                       "         an alloc/free accounting bug: cbBefore=%d cbAfter=%d\n", cbBefore, cbAfter);
-        RTHeapOffsetDump(Heap, (PFNRTHEAPOFFSETPRINTF)RTPrintf);
+        RTHeapOffsetDump(Heap, (PFNRTHEAPOFFSETPRINTF)(uintptr_t)RTPrintf);
     }
 
     /* relocate and free the bits in heap2 now. */
